requirements.txt
alembic==1.13.2
bcrypt==4.2.0
bleach==6.1.0
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-Limiter==3.8.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
Flask-WTF==1.2.1
greenlet==3.0.3
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
limits==3.13.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
Pygments==2.18.0
python-dotenv==1.0.1
rich==13.8.1
six==1.16.0
SQLAlchemy==2.0.32
typing_extensions==4.12.2
webencodings==0.5.1
Werkzeug==3.0.3
wrapt==1.16.0
WTForms==3.1.2
